changes in different schema, return type and exception message
*different services id is changed to global id
*different services rating changed to global rating
*return type change to ResponseEntity with proper status
*data fetch of hotel in terms of name or location api done
*Exceptional handling message updated
*User schema primary key changed

+++ src/main/java/com/example/models/RatingInventory.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class RatingInventory {
 
-    private float quality=1;
-    private float asAdvertised=1;
-    private float satisfaction=1;
+    private Double quality = 1.0;
+    private Double asAdvertised = 1.0;
+    private Double satisfaction = 1.0;
 
 }
